--- erukar/server/components/__init__.py
@@ -1,4 +1,0 @@
-from .DataAccess import DataAccess
-from .Interface import Interface
-
-__all__ = ['DataAccess', 'Interface']
